Rename i-google-analytics-command.ts
