+++ projects/ngx-google-analytics/src/lib/interfaces/i-google-analytics-command.ts
